Add files via upload
Here is a list of CAN Bus IDs I was able to sniff of a Panamera 970 cluster using a custom made Arduino sniffer code via Seeed Can bus shield and Arduino Mega 2560. The main idea for sniffing it off was to interface it with simhub software to make the cluster function with gaming simulators.